Added classDiagram

+++ PECourseProject/Point.cpp
@@ -5,10 +5,10 @@
 Point::Point(): x(0), y(0)
 {}
 
-Point::Point(int x, int y) :x(x), y(y)
+Point::Point(float x, float y) :x(x), y(y)
 {}
 
-double Point::distanceTo(Point other)
+float Point::distanceTo(Point other)
 {
 	return sqrt(pow(other.x - this->x, 2) + pow(other.y - this->y, 2));
 }
